Port HCenter element to new trait
This includes changing the width constraint from an Option to a struct with a float and a bool. Among other things this makes it possible for HCenter to work on long lines of text that would otherwise overflow.
This would also have been possible in the previous model, but it would have required measuring the content twice in the HCenter measure implementation.