Change MinFirstHeight to not grow it's height to min_first_height
Instead it now measures to see if it would break if first_height is less
than min_first_height. If the inner element does not break in that case
we also do not pre break.

If we didn't do anything like this the issue would be that the element
could violate assumtions. It could not just be drawn with the size it
returned in measure because it might then trigger a pre-break.