Fix background drawing in StyledBox element
The last background box that got drawn used the location of the last
break that was performed instead of the last location that was ever
broken to.

This also mitigates an issue where the Page element always adds a layer
ontop after each break, even when repeating breaks. That means any child
element of Page that repeats a break will get the top layer after each
break, which means it can't draw a background on the layer returned
there that is below other content that was drawn before.